エンジン情報マネージャーとエンジンプロセスマネージャーを分ける
