+++ src/backend/electron/manager/engineProcessManager.ts
@@ -2,7 +2,6 @@ import { spawn, ChildProcess } from "child_process";
 import path from "path";
 import fs from "fs";
 import treeKill from "tree-kill";
-import shlex from "shlex";
 
 import { app, dialog } from "electron"; // FIXME: ここでelectronをimportするのは良くない
 
@@ -13,7 +12,7 @@ import {
   getProcessNameFromPid,
   isAssignablePort,
   url2HostInfo,
-} from "./portManager";
+} from "../portHelper";
 
 import {
   EngineInfo,
@@ -21,7 +20,7 @@ import {
   MinimumEngineManifestType,
   EngineId,
   minimumEngineManifestSchema,
-  envEngineInfoSchema,
+  EngineSettings,
 } from "@/type/preload";
 import { AltPortInfos } from "@/store/type";
 import { BaseConfigManager } from "@/backend/common/ConfigManager";
@@ -31,180 +30,53 @@ type EngineProcessContainer = {
   engineProcess?: ChildProcess;
 };
 
-/**
- * デフォルトエンジンの情報を作成する
- */
-function createDefaultEngineInfos(defaultEngineDir: string): EngineInfo[] {
-  // TODO: envから直接ではなく、envに書いたengine_manifest.jsonから情報を得るようにする
-  const defaultEngineInfosEnv =
-    import.meta.env.VITE_DEFAULT_ENGINE_INFOS ?? "[]";
-
-  const envSchema = envEngineInfoSchema.array();
-  const engines = envSchema.parse(JSON.parse(defaultEngineInfosEnv));
-
-  return engines.map((engineInfo) => {
-    return {
-      ...engineInfo,
-      isDefault: true,
-      type: "path",
-      executionFilePath: path.resolve(engineInfo.executionFilePath),
-      path:
-        engineInfo.path == undefined
-          ? undefined
-          : path.resolve(defaultEngineDir, engineInfo.path),
-    } satisfies EngineInfo;
-  });
-}
-
-export class EngineManager {
+/** エンジンプロセスを管理するクラス */
+export class EngineProcessManager {
   configManager: BaseConfigManager;
-  defaultEngineDir: string;
   vvppEngineDir: string;
   onEngineProcessError: (engineInfo: EngineInfo, error: Error) => void;
+  engineInfosFetcher: () => Readonly<EngineInfo>[];
+  engineInfoPortChanger: (engineId: EngineId, port: number) => void;
+  engineSettingsFetcher: () => Partial<EngineSettings>;
 
   defaultEngineInfos: EngineInfo[] = [];
   additionalEngineInfos: EngineInfo[] = [];
-  engineProcessContainers: Record<EngineId, EngineProcessContainer>;
+  engineProcessContainers: Record<EngineId, EngineProcessContainer> = {};
 
-  public altPortInfo: AltPortInfos = {};
+  /** 代替ポート情報 */
+  public altPortInfos: AltPortInfos = {};
 
-  constructor({
-    configManager,
-    defaultEngineDir,
-    vvppEngineDir,
-    onEngineProcessError,
-  }: {
+  constructor(payload: {
     configManager: BaseConfigManager;
-    defaultEngineDir: string;
     vvppEngineDir: string;
     onEngineProcessError: (engineInfo: EngineInfo, error: Error) => void;
+    /** エンジン情報を取得する関数 */
+    engineInfosFetcher: () => Readonly<EngineInfo>[];
+    /** エンジン情報のhost内のportを置き換える関数 */
+    engineInfoPortChanger: (engineId: EngineId, port: number) => void;
+    /** エンジン設定を取得する関数 */
+    engineSettingsFetcher: () => Partial<EngineSettings>;
   }) {
-    this.configManager = configManager;
-    this.defaultEngineDir = defaultEngineDir;
-    this.vvppEngineDir = vvppEngineDir;
-    this.onEngineProcessError = onEngineProcessError;
-    this.engineProcessContainers = {};
+    this.configManager = payload.configManager;
+    this.vvppEngineDir = payload.vvppEngineDir;
+    this.onEngineProcessError = payload.onEngineProcessError;
+    this.engineInfosFetcher = payload.engineInfosFetcher;
+    this.engineInfoPortChanger = payload.engineInfoPortChanger;
+    this.engineSettingsFetcher = payload.engineSettingsFetcher;
   }
 
   /**
-   * 追加エンジンの一覧を作成する。
-   * FIXME: store.get("registeredEngineDirs")への副作用をEngineManager外に移動する
+   * 代替ポートの情報を初期化する。
    */
-  private createAdditionalEngineInfos(): EngineInfo[] {
-    const engines: EngineInfo[] = [];
-    const addEngine = (engineDir: string, type: "vvpp" | "path") => {
-      const manifestPath = path.join(engineDir, "engine_manifest.json");
-      if (!fs.existsSync(manifestPath)) {
-        return "manifestNotFound";
-      }
-      let manifest: MinimumEngineManifestType;
-      try {
-        manifest = minimumEngineManifestSchema.parse(
-          JSON.parse(fs.readFileSync(manifestPath, { encoding: "utf8" })),
-        );
-      } catch (e) {
-        return "manifestParseError";
-      }
-
-      const [command, ...args] = shlex.split(manifest.command);
-
-      engines.push({
-        uuid: manifest.uuid,
-        host: `http://127.0.0.1:${manifest.port}`,
-        name: manifest.name,
-        path: engineDir,
-        executionEnabled: true,
-        executionFilePath: path.join(engineDir, command),
-        executionArgs: args,
-        type,
-        isDefault: false,
-      } satisfies EngineInfo);
-      return "ok";
-    };
-    for (const dirName of fs.readdirSync(this.vvppEngineDir)) {
-      const engineDir = path.join(this.vvppEngineDir, dirName);
-      if (!fs.statSync(engineDir).isDirectory()) {
-        log.log(`${engineDir} is not directory`);
-        continue;
-      }
-      if (dirName === ".tmp") {
-        continue;
-      }
-      const result = addEngine(engineDir, "vvpp");
-      if (result !== "ok") {
-        log.log(`Failed to load engine: ${result}, ${engineDir}`);
-      }
-    }
-    // FIXME: この関数の引数でregisteredEngineDirsを受け取り、動かないエンジンをreturnして、EngineManager外でconfig.setする
-    for (const engineDir of this.configManager.get("registeredEngineDirs")) {
-      const result = addEngine(engineDir, "path");
-      if (result !== "ok") {
-        log.log(`Failed to load engine: ${result}, ${engineDir}`);
-        // 動かないエンジンは追加できないので削除
-        // FIXME: エンジン管理UIで削除可能にする
-        dialog.showErrorBox(
-          "エンジンの読み込みに失敗しました。",
-          `${engineDir}を読み込めませんでした。このエンジンは削除されます。`,
-        );
-        this.configManager.set(
-          "registeredEngineDirs",
-          this.configManager
-            .get("registeredEngineDirs")
-            .filter((p) => p !== engineDir),
-        );
-      }
-    }
-    return engines;
-  }
-
-  /**
-   * 全てのエンジンの一覧を取得する。デフォルトエンジン＋追加エンジン。
-   */
-  fetchEngineInfos(): EngineInfo[] {
-    return [...this.defaultEngineInfos, ...this.additionalEngineInfos];
-  }
-
-  /**
-   * エンジンの情報を取得する。存在しない場合はエラーを返す。
-   */
-  fetchEngineInfo(engineId: EngineId): EngineInfo {
-    const engineInfos = this.fetchEngineInfos();
-    const engineInfo = engineInfos.find(
-      (engineInfo) => engineInfo.uuid === engineId,
-    );
-    if (!engineInfo) {
-      throw new Error(`No such engineInfo registered: engineId == ${engineId}`);
-    }
-    return engineInfo;
-  }
-
-  /**
-   * エンジンのディレクトリを取得する。存在しない場合はエラーを返す。
-   */
-  fetchEngineDirectory(engineId: EngineId): string {
-    const engineInfo = this.fetchEngineInfo(engineId);
-    const engineDirectory = engineInfo.path;
-    if (engineDirectory == undefined) {
-      throw new Error(`engineDirectory is undefined: engineId == ${engineId}`);
-    }
-
-    return engineDirectory;
-  }
-
-  /**
-   * EngineInfosとAltPortInfoを初期化する。
-   */
-  initializeEngineInfosAndAltPortInfo() {
-    this.defaultEngineInfos = createDefaultEngineInfos(this.defaultEngineDir);
-    this.additionalEngineInfos = this.createAdditionalEngineInfos();
-    this.altPortInfo = {};
+  initializeAltPortInfos() {
+    this.altPortInfos = {};
   }
 
   /**
    * 全てのエンジンを起動する。
    */
   async runEngineAll() {
-    const engineInfos = this.fetchEngineInfos();
+    const engineInfos = this.engineInfosFetcher();
     log.info(`Starting ${engineInfos.length} engine/s...`);
 
     for (const engineInfo of engineInfos) {
@@ -217,7 +89,7 @@ export class EngineManager {
    * エンジンを起動する。
    */
   async runEngine(engineId: EngineId) {
-    const engineInfos = this.fetchEngineInfos();
+    const engineInfos = this.engineInfosFetcher();
     const engineInfo = engineInfos.find(
       (engineInfo) => engineInfo.uuid === engineId,
     );
@@ -278,14 +150,14 @@ export class EngineManager {
         throw new Error("No Alternative Port Found");
       }
 
-      // 代替ポートの情報
-      this.altPortInfo[engineId] = {
+      // 代替ポート情報を更新
+      this.altPortInfos[engineId] = {
         from: engineHostInfo.port,
         to: altPort,
       };
 
-      // 代替ポートを設定
-      engineInfo.host = `${engineHostInfo.protocol}//${engineHostInfo.hostname}:${altPort}`;
+      // エンジン情報に代替ポートを反映
+      this.engineInfoPortChanger(engineId, altPort);
       log.warn(
         `ENGINE ${engineId}: Applied Alternative Port: ${engineHostInfo.port} -> ${altPort}`,
       );
@@ -544,7 +416,7 @@ export class EngineManager {
       return "invalidManifest";
     }
 
-    const engineInfos = this.fetchEngineInfos();
+    const engineInfos = this.engineInfosFetcher();
     if (
       engineInfos.some((engineInfo) => engineInfo.uuid === manifestContent.uuid)
     ) {
@@ -554,4 +426,4 @@ export class EngineManager {
   }
 }
 
-export default EngineManager;
+export default EngineProcessManager;
